migrating useState Lesson into shared endpoint
