+++ pages/api/studentPortal/lessons/web/react/index.ts
@@ -3,6 +3,7 @@ import { getGithubLessonData } from "./github";
 import { getPropsLessonData } from "./props";
 import { getTailwindColourStylingLessonData } from "./tailwindcss-colourstyling";
 import { getTailwindGridFlexLessonData } from "./tailwindcss-gridflex";
+import { getUseStateLessonData } from "./useState";
 
 export function getLessonForReactCourse(lessonId: string) {
   if (lessonId === "components") {
@@ -15,6 +16,8 @@ export function getLessonForReactCourse(lessonId: string) {
     return getTailwindColourStylingLessonData();
   } else if (lessonId === "tailwindcss-gridflex") {
     return getTailwindGridFlexLessonData();
+  } else if (lessonId === "useState") {
+    return getUseStateLessonData();
   }
 
   throw new Error(`Could not find lessons`);
